FSM-11 Add react router
Added a jsconfig file so the baseurl is treated as `src`
Added react helmet and react router dom packages
Added redirects file for netlify to do with SPA router issues
Created routes for map + list view
