+++ jsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es2016",
+    "jsx": "preserve",
+    "checkJs": true,
+    "baseUrl": "src"
+  },
+  "exclude": ["node_modules", "**/node_modules/*"]
+}
