+++ src/services/history.js
@@ -1,0 +1,3 @@
+import { createBrowserHistory } from "history";
+const history = createBrowserHistory();
+export default history;
